+++ LightBulb/NetworkTopology/AbstractNetworkTopology.hpp
@@ -74,15 +74,15 @@ public:
 
 	virtual double getNetInput(int layerIndex, int neuronIndex) = 0;
 
-	virtual Eigen::VectorXf getNetInputVector(int layerIndex) = 0;
+	virtual Eigen::VectorXd getNetInputVector(int layerIndex) = 0;
 
-	virtual Eigen::VectorXf getEfferentWeightVector(int layerIndex, int neuronIndex) = 0;
+	virtual Eigen::VectorXd getEfferentWeightVector(int layerIndex, int neuronIndex) = 0;
 
-	virtual Eigen::VectorXf getActivationVector(int layerIndex) = 0;
+	virtual Eigen::VectorXd getActivationVector(int layerIndex) = 0;
 
-	virtual Eigen::MatrixXf getAfferentWeightsPerLayer(int layerIndex) = 0;
+	virtual Eigen::MatrixXd getAfferentWeightsPerLayer(int layerIndex) = 0;
 
-	virtual void setAfferentWeightsPerLayer(int layerIndex, Eigen::MatrixXf& newWeights) = 0;
+	virtual void setAfferentWeightsPerLayer(int layerIndex, Eigen::MatrixXd& newWeights) = 0;
 };
 
 #endif
