+++ LightBulb/NetworkTopology/LayeredNetwork.hpp
@@ -142,17 +142,17 @@ public:
 
 	double getNetInput(int layerIndex, int neuronIndex);
 
-	Eigen::VectorXf getNetInputVector(int layerIndex);
+	Eigen::VectorXd getNetInputVector(int layerIndex);
 
-	Eigen::VectorXf getEfferentWeightVector(int layerIndex, int neuronIndex);
+	Eigen::VectorXd getEfferentWeightVector(int layerIndex, int neuronIndex);
 
 	AbstractActivationFunction* getInnerActivationFunction();
 
-	Eigen::VectorXf getActivationVector(int layerIndex);
+	Eigen::VectorXd getActivationVector(int layerIndex);
 
-	Eigen::MatrixXf getAfferentWeightsPerLayer(int layerIndex);
+	Eigen::MatrixXd getAfferentWeightsPerLayer(int layerIndex);
 
-	void setAfferentWeightsPerLayer(int layerIndex, Eigen::MatrixXf& newWeights);
+	void setAfferentWeightsPerLayer(int layerIndex, Eigen::MatrixXd& newWeights);
 };
 
 #endif
